Forgot to check if token in SessionStorage
If the user does not check off the 'Keep me logged in' checkbox then we
store the token in the sessionStorage. This way when he closes the tab
or browser it removes the token. But when we need to use the token on
the webapp we are never checking the SessionStorage.